fix: wrong scope for reactive block statement
