+++ tests/fixtures/parser/ast/reactive-with-var03-input.svelte
@@ -1,0 +1,8 @@
+<script>
+    let foo
+    $: {
+        foo = 1
+    }
+</script>
+
+{foo}
