+++ tests/fixtures/parser/ast/reactive-with-var04-ts-input.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+    let foo: number
+    $: {
+        foo = 1
+    }
+</script>
+
+{foo}
